chore(menu): demo拆解与规范
